% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_testing.R
\name{chi_squared_conditions_test}
\alias{chi_squared_conditions_test}
\title{Test the Conditions for the Chi-Squared Test}
\usage{
chi_squared_conditions_test(var_1, var_2)
}
\arguments{
\item{var_1}{the column name of the first variable in the dataset.}

\item{var_2}{the column name of the second variable in the dataset.}
}
\value{
A list:
\itemize{
\item \code{verdict}: A boolean that is set to \code{TRUE} if the conditions for the test have been met.
\item \code{message}: A character string describing why the conditions for the test have not been met or
signaling that they have been met.
\item \code{contingency_table}: The contingency table used for the test.
\item \code{independent_table}: The frequency table under the assumption that the variables are independent.
}
}
\description{
Test the Conditions for the Chi-Squared Test
}
\examples{
chi_squared_conditions_test("Fun with friends", "Politics")$message
}
