% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_testing.R
\name{chi_squared_test}
\alias{chi_squared_test}
\title{Chi-Squared Test}
\usage{
chi_squared_test(var_1, var_2)
}
\arguments{
\item{var_1}{the column name of the first variable in the dataset.}

\item{var_2}{the column name of the second variable in the dataset.}
}
\value{
A list that contains the results of the test:
\itemize{
\item \code{statistic}: The Chi-squared statistic.
\item \code{chisq_df}: The degrees of freedom.
\item \code{p_value}: The p-value.
}
}
\description{
Conducts a Chi-squared test on the variables \code{var_1} and \code{var_2}.
}
\examples{
chi_squared_test("Pets", "Dancing")
}
