% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_representation.R
\name{combine_bar_charts}
\alias{combine_bar_charts}
\title{Combine Bar Charts}
\usage{
combine_bar_charts(...)
}
\arguments{
\item{...}{A sequence of bar charts.}
}
\value{
A grid of bar charts arranged row by row.
}
\description{
Combine Bar Charts
}
\examples{
p1 <- var_bar_chart("Mathematics", "Biology")
p2 <-  var_bar_chart("Mathematics", "Shopping")
p3 <- var_bar_chart("Mathematics", "Pets")
combine_bar_charts(p1, p2, p3)
}
