% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_representation.R
\name{combine_pie_charts}
\alias{combine_pie_charts}
\title{Combine Pie Charts}
\usage{
combine_pie_charts(...)
}
\arguments{
\item{...}{A sequence of pie charts.}
}
\value{
A grid of pie charts.
}
\description{
Combine Pie Charts
}
\examples{
p1 <- one_var_pie_chart("Art exhibitions")
p2 <- one_var_pie_chart("Chemistry")
p3 <- one_var_pie_chart("Dancing")
combine_pie_charts(p1, p2, p3)
}
