% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{get_data}
\alias{get_data}
\title{Get the Data}
\usage{
get_data()
}
\value{
A data frame: the dataset used in this package.
}
\description{
This function loads the dataset used in this package.
}
\examples{
get_data()
}
