% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{get_one_var_data}
\alias{get_one_var_data}
\title{Get Data for a Variable}
\usage{
get_one_var_data(var)
}
\arguments{
\item{var}{the column name of the variable in the dataset.}
}
\value{
A data frame with one column: \code{col}. It contains entries of the form
"Very/Fairly/Not much interested in <column_name_of_variable>". These entries correspond to values 5/4-3/2-1
in the original dataset.
}
\description{
Returns the data for the analysis of a column in the dataset.
The returned data excludes NA entries.
}
\examples{
my_var_data <- get_one_var_data("Psychology")
table(my_var_data$col)
}
