% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{get_var_data}
\alias{get_var_data}
\title{Get Data for Two Variables}
\usage{
get_var_data(var_1, var_2)
}
\arguments{
\item{var_1}{the column name of the first variable in the dataset.}

\item{var_2}{the column name of the second variable in the dataset.}
}
\value{
A data frame with two columns: "col_1" and "col_2". For i = 1, 2, the column
"col_i" contains entries of the form "Very/Fairly/Not much interested in <var_i>".
These entries correspond to values 5/4-3/2-1 in the original dataset.
}
\description{
Returns the data for the analysis of two columns in the dataset.
}
\examples{
my_var_data <- get_var_data("Mathematics", "Biology")
table(my_var_data$col_1, my_var_data$col_2)
}
