% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{get_var_names}
\alias{get_var_names}
\title{Variable Names}
\usage{
get_var_names()
}
\value{
A character vector: the column names of the dataset used in the package.
}
\description{
This function returns the names of the variables (columns) in the dataset used in the package.
}
\examples{
get_var_names()
}
