% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_representation.R
\name{one_var_freq_table}
\alias{one_var_freq_table}
\title{Frequency Table for a Variable}
\usage{
one_var_freq_table(var)
}
\arguments{
\item{var}{the column name of the variable in the dataset.}
}
\value{
The frequency table for the variable, excluding NA entries.
}
\description{
Returns the frequency table for a variable, excluding NA entries.
}
\examples{
one_var_freq_table("Shopping")
}
