% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_representation.R
\name{one_var_pie_chart}
\alias{one_var_pie_chart}
\title{Plot a Pie Chart for a Variable}
\usage{
one_var_pie_chart(var)
}
\arguments{
\item{var}{the column name of the variable in the dataset.}
}
\value{
A pie chart for the variable.
}
\description{
Plot a Pie Chart for a Variable
}
\examples{
one_var_pie_chart("Celebrities")
}
