% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_representation.R
\name{var_bar_chart}
\alias{var_bar_chart}
\title{Plot a Bar Chart for Two Variables}
\usage{
var_bar_chart(var_x, var_y)
}
\arguments{
\item{x_var}{the column name of the variable in the dataset that is used for the x-axis.}

\item{y_var}{the column name of the variable in the dataset that is used for the y-axis.}
}
\value{
A bar chart for the variables.
}
\description{
Plot a Bar Chart for Two Variables
}
\examples{
var_bar_chart("History", "Science and technology")
}
