% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_testing.R
\name{var_extreme_chisq_stat_vals}
\alias{var_extreme_chisq_stat_vals}
\title{Extreme Chi-Squared Statistic Values of a Variable}
\usage{
var_extreme_chisq_stat_vals(var, M = 3, biggest_chisq_vals = FALSE)
}
\arguments{
\item{var}{the column name of the variable in the dataset.}

\item{M}{the number of extreme values to display.}

\item{biggest_chisq_vals}{whether to show the biggest Chi-squared statistics (\code{TRUE}) or
the smallest (\code{FALSE}).}
}
\value{
A data frame with columns \code{var_name} and \code{chi_sq_val}. The data frame
includes the variables that give the smallest or biggest Chi-squared statistic values
when conducting the Chi-squared test with the variable \code{var}.
}
\description{
Helps to discover which variables give the smallest or biggest Chi-squared statistic values
when conducting the Chi-squared test with the variable \code{var}.
Variables are included in the results only if they meet the conditions for the Chi-squared test.
}
\examples{
var_extreme_chisq_stat_vals("Mathematics", M = 5, biggest_chisq_vals = TRUE)
}
