% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_representation.R
\name{var_freq_table}
\alias{var_freq_table}
\title{Frequency Table for Two Variables}
\usage{
var_freq_table(var_1, var_2)
}
\arguments{
\item{var_1}{the column name of the first variable in the dataset.}

\item{var_2}{the column name of the second variable in the dataset.}
}
\value{
A frequency table for the two variables, excluding NA entries.
}
\description{
Returns the frequency table for two variables, excluding NA entries.
}
\examples{
var_freq_table("Adrenaline sports", "Cars")
}
