% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_testing.R
\name{visualize_chi_squared_test}
\alias{visualize_chi_squared_test}
\title{Visualize the Chi-Squared Test}
\usage{
visualize_chi_squared_test(var_1, var_2)
}
\arguments{
\item{var_1}{the column name of the first variable in the dataset.}

\item{var_2}{the column name of the second variable in the dataset.}
}
\value{
A plot that visualizes the result of the test.
}
\description{
This function conducts a Chi-squared test on the variables var_1 and var_2.
It does so by displaying two generated null distributions: one using a theoretical approximation
(shown by the continuous line) and another using resampling methods (shown by the bars). Finally,
it draws a vertical red line, representing the observed statistic of the test.
}
\examples{
visualize_chi_squared_test("Mathematics", "Biology")
}
