% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_document.R
\name{sample_function}
\alias{sample_function}
\title{Sample Function}
\usage{
sample_function(x = 5)
}
\arguments{
\item{x}{An integer}
}
\value{
An almost random number.
}
\description{
Sample Function
}
\examples{
sample_function()
sample_function(3)
}
